+++ l4-proxy/l4-proxy.go
@@ -5,14 +5,12 @@ import (
     "encoding/json"
     "flag"
     "fmt"
-    "io"
     "io/ioutil"
     "log"
     "net"
     "net/http"
     "os"
     "strings"
-    "syscall"
     "time"
 
     "github.com/libp2p/go-libp2p-core/peer"
@@ -57,233 +55,6 @@ type Forwarder struct {
 
 // Maps a remote addr to existing ServiceProxy for that addr
 var serv2Fwd = make(map[string]Forwarder)
-
-// Data forwarder
-// Simplified version of pipe() from https://github.com/jpillora/go-tcp-proxy/blob/master/proxy.go
-func fwdData(src, dst net.Conn) {
-    buf := make([]byte, 0xffff) // 64k buffer
-    for {
-        // NOTE: Using io.Copy or io.CopyBuffer slows down *a lot* after 2-3 runs.
-        //       Not clear why right now. Thus, do the copy manually.
-        nBytes, err := src.Read(buf)
-        if err != nil {
-            if err != io.EOF {
-                log.Printf("ERROR: Unable to read from connection %s\n", src)
-            }
-            return
-        }
-        data := buf[:nBytes]
-
-        nBytes, err = dst.Write(data) // TODO: Ensure nBytes is written
-        if err != nil {
-            if err != io.EOF {
-                fmt.Printf("ERROR: Unable to write to connection %s\n", dst)
-            }
-            return
-        }
-    }
-}
-
-// Connection forwarder
-func fwdLocalConn(lConn net.Conn, targetAddr string) {
-    defer lConn.Close()
-    log.Printf("Accepted conn: %s\n", lConn)
-
-    // Resolve and open connection to destination service
-    rAddr, err := net.ResolveTCPAddr("tcp", targetAddr)
-    if err != nil {
-        log.Printf("ERROR: Unable to resolve target address %s\n%v\n", targetAddr, err)
-        return
-    }
-
-    rConn, err := net.DialTCP("tcp", nil, rAddr)
-    if err != nil {
-        log.Printf("ERROR: Unable to dial target address %s\n%v\n", targetAddr, err)
-        return
-    }
-    defer rConn.Close()
-
-    // Forward data in each direction
-    go fwdData(lConn, rConn)
-    fwdData(rConn, lConn)
-}
-
-// Implementation of ServiceProxy
-func tcpServiceProxy(listen net.Listener, targetAddr string) {
-    defer listen.Close()
-
-    for {
-        conn, err := listen.Accept()
-        if err != nil {
-            if err == syscall.EINVAL {
-                log.Printf("Listener (%s) closed", listen.Addr())
-                break
-            }
-            log.Printf("Listener (%s) unable to accept connection\n%v\n", listen.Addr(), err)
-            continue
-        }
-
-        // Forward connection
-        go fwdLocalConn(conn, targetAddr)
-    }
-}
-
-// UDP data forwarder
-func udpFwdData(src, dst *net.UDPConn, dstAddr net.Addr) {
-    buf := make([]byte, 0xffff) // 64k buffer
-    for {
-        nBytes, err := src.Read(buf)
-        if err != nil {
-            if err == syscall.EINVAL {
-                log.Printf("Connection %s <=> %s closed", src.LocalAddr(), src.RemoteAddr())
-                break
-            } else {
-                log.Printf("ERROR: Unable to read from UDP connection %s <=> %s\n%v\n",
-                    src.LocalAddr(), src.RemoteAddr(), err)
-            }
-            continue
-        }
-        data := buf[:nBytes]
-
-        nBytes, err = dst.WriteTo(data, dstAddr) // TODO: Ensure nBytes actually written
-        if err != nil {
-            if err == syscall.EINVAL {
-                log.Printf("Connection %s <=> %s closed", src.LocalAddr(), src.RemoteAddr())
-                break
-            } else {
-                log.Printf("ERROR: Unable to write to UDP connection %s <=> %s\n%v\n",
-                    src.LocalAddr(), src.RemoteAddr(), err)
-            }
-            continue
-        }
-    }
-}
-
-// Demultiplex incoming packets and forward them to destination
-func udpServiceProxy(lConn *net.UDPConn, targetAddr string) {
-    defer lConn.Close()
-
-    var err error
-
-    // Resolve and open connection to destination service
-    rAddr, err := net.ResolveUDPAddr("udp", targetAddr)
-    if err != nil {
-        log.Printf("ERROR: Unable to resolve UDP target address %s\n%v\n", targetAddr, err)
-        return
-    }
-
-    // Since there's no per-client UDP connection object, we'll need to do some
-    // manual demultiplexing. Create a separate outgoing UDP "connection" to the
-    // same target service per unique client.
-    client2RConn := make(map[net.Addr]*net.UDPConn)
-
-    var rConn *net.UDPConn
-    var exists bool
-    var from net.Addr
-    var nBytes int
-    buf := make([]byte, 0xffff) // 64k buffer
-    for {
-        nBytes, from, err = lConn.ReadFrom(buf)
-        if err != nil {
-            if err == syscall.EINVAL {
-                log.Printf("Connection %s <=> %s closed",
-                    lConn.LocalAddr(), lConn.RemoteAddr())
-            } else {
-                log.Printf("ERROR: Unable to read from UDP connection %s <=> %s\n%v\n",
-                    lConn.LocalAddr(), lConn.RemoteAddr(), err)
-            }
-            return
-        }
-
-        // Create per-client connection with remote service
-        if rConn, exists = client2RConn[from]; !exists {
-            rConn, err = net.DialUDP("udp", nil, rAddr)
-            if err != nil {
-                log.Printf("ERROR: Unable to dial UDP target address %s\n%v\n",
-                    targetAddr, err)
-                continue // Or return? Other connections may be okay...
-            }
-            defer rConn.Close()
-
-            client2RConn[from] = rConn
-
-            // Create separate goroutine to handle reverse path
-            go udpFwdData(rConn, lConn, from)
-        }
-
-        data := buf[:nBytes]
-
-        nBytes, err = rConn.Write(data)
-        if err != nil {
-            if err == syscall.EINVAL {
-                log.Printf("Connection %s <=> %s closed",
-                    rConn.LocalAddr(), rConn.RemoteAddr())
-            } else {
-                log.Printf("ERROR: Unable to write to UDP connection %s <=> %s\n%v\n",
-                    rConn.LocalAddr(), rConn.RemoteAddr(), err)
-            }
-            continue
-        }
-
-    }
-}
-
-// Open TCP tunnel to service and open local TCP listening port
-// NOTE: Currently doesn't support SSL/TLS
-//
-// TODO: In the future, perhaps the connection type info (TCP/UDP) can be stored in hash-lookup?
-// TODO: In the future, move towards proxy-to-proxy implementation
-//       This may be useful? https://github.com/libp2p/go-libp2p-gostream
-// Start separate goroutine for handling connections and proxying to service
-//
-// Returns a string of the new TCP listening endpoint address
-func openTCPProxy(serviceAddr string) (string, error) {
-    var listenAddr string
-    if _, exists := serv2Fwd[serviceAddr]; !exists {
-        listen, err := net.Listen("tcp", ctrlHost + ":") // automatically choose port
-        if err != nil {
-            return "", fmt.Errorf("Unable to open TCP listening port\n")
-        }
-
-        listenAddr = listen.Addr().String()
-        serv2Fwd[serviceAddr] = Forwarder {
-            ListenAddr: listenAddr,
-            tcpWorker: tcpServiceProxy,
-        }
-        go serv2Fwd[serviceAddr].tcpWorker(listen, serviceAddr)
-    } else {
-        listenAddr = serv2Fwd[serviceAddr].ListenAddr
-    }
-
-    return listenAddr, nil
-}
-
-// Open UDP tunnel to service and open local UDP listening port
-func openUDPProxy(serviceAddr string) (string, error) {
-    var listenAddr string
-    if _, exists := serv2Fwd[serviceAddr]; !exists {
-        udpAddr, err := net.ResolveUDPAddr("udp", ctrlHost + ":") // choose port
-        if err != nil {
-            return "", fmt.Errorf("Unable to resolve UDP address\n%w\n", err)
-        }
-
-        conn, err := net.ListenUDP("udp", udpAddr)
-        if err != nil {
-            return "", fmt.Errorf("Unable to open UDP listening port\n")
-        }
-
-        listenAddr = conn.LocalAddr().String()
-        serv2Fwd[serviceAddr] = Forwarder {
-            ListenAddr: listenAddr,
-            udpWorker: udpServiceProxy,
-        }
-        go serv2Fwd[serviceAddr].udpWorker(conn, serviceAddr)
-    } else {
-        listenAddr = serv2Fwd[serviceAddr].ListenAddr
-    }
-
-    return listenAddr, nil
-}
 
 // Handles the "proxying" part of proxy
 // TODO: Refactor this function
